build: 升级 Spring Boot 2.3.12→3.4.5 / Java 8→21,配置指向新环境
M1·Phase A+B 地基改造,为后续 Spring AI(要求 Boot3+JDK17)铺路:
- 依赖换代:javax→jakarta;MyBatis-Plus 换 boot3 starter 3.5.9 + jsqlparser
  模块;MySQL 驱动 com.mysql.cj.jdbc.Driver;Redisson 3.13.6→3.34.1;
  去掉手钉的 lettuce/spring-data-redis 版本交给 BOM 管理。
- 代码适配:count() 返回 Long(2 处);BlogServiceImpl 自注入加 @Lazy
  (Boot3 默认禁循环依赖)。
- 配置:数据源指向 MySQL 3308/库 ai_customer_service;Redis 前缀迁
  spring.data.redis 指向 6381;RedissonConfig 同步 6381。
- 验证:package -DskipTests 通过;java -jar 对接 Docker 成功启动并响应 HTTP。

+++ backend/src/main/java/com/hmdp/service/impl/FollowServiceImpl.java
@@ -15,7 +15,7 @@ import com.hmdp.utils.UserHolder;
 import org.springframework.data.redis.core.StringRedisTemplate;
 import org.springframework.stereotype.Service;
 
-import javax.annotation.Resource;
+import jakarta.annotation.Resource;
 import java.util.List;
 import java.util.Set;
 import java.util.stream.Collectors;
@@ -63,7 +63,7 @@ public class FollowServiceImpl extends ServiceImpl<FollowMapper, Follow> impleme
     @Override
     public Result followOrNot(Long followUserId) {
         Long userId = UserHolder.getUser().getId();
-        Integer count = query().eq("user_id", userId).eq("follow_user_id", followUserId).count();
+        Long count = query().eq("user_id", userId).eq("follow_user_id", followUserId).count();
         return Result.ok(count > 0);
     }
